Make endpoints secure

+++ HRPotter/Controllers/CompaniesController.cs
@@ -11,6 +11,7 @@ using static HRPotter.Controllers.UsersController;
 
 namespace HRPotter.Controllers
 {
+    [ProducesResponseType(StatusCodes.Status403Forbidden)]
     [Route("[controller]")]
     [Authorize]
     public class CompaniesController : Controller
@@ -36,6 +37,11 @@ namespace HRPotter.Controllers
         [HttpGet]
         public async Task<IActionResult> Index()
         {
+            if (HRPotterUser.Role != "Admin")
+            {
+                return Forbid();
+            }
+
             return View(await _context.Companies.ToListAsync());
         }
 
@@ -46,6 +52,11 @@ namespace HRPotter.Controllers
         [HttpGet]
         public async Task<IActionResult> GetCompaniesTable()
         {
+            if (HRPotterUser.Role != "Admin")
+            {
+                return Forbid();
+            }
+
             return PartialView("_CompaniesTable", await _context.Companies.ToListAsync());
         }
 
@@ -60,6 +71,11 @@ namespace HRPotter.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create(Company company)
         {
+            if (HRPotterUser.Role != "Admin")
+            {
+                return Forbid();
+            }
+
             if (ModelState.IsValid)
             {
                 _context.Add(company);
@@ -81,6 +97,11 @@ namespace HRPotter.Controllers
         [HttpGet]
         public async Task<IActionResult> Edit(int? id)
         {
+            if (HRPotterUser.Role != "Admin")
+            {
+                return Forbid();
+            }
+
             if (id == null)
             {
                 return NotFound();
@@ -106,6 +127,11 @@ namespace HRPotter.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("Id,Name")] Company company)
         {
+            if (HRPotterUser.Role != "Admin")
+            {
+                return Forbid();
+            }
+
             if (company == null || id != company.Id)
             {
                 return NotFound();
@@ -145,6 +171,11 @@ namespace HRPotter.Controllers
         [HttpGet]
         public async Task<IActionResult> GetDeleteModal(int? id)
         {
+            if (HRPotterUser.Role != "Admin")
+            {
+                return Forbid();
+            }
+
             if (id == null)
             {
                 return NotFound();
@@ -170,6 +201,11 @@ namespace HRPotter.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            if (HRPotterUser.Role != "Admin")
+            {
+                return Forbid();
+            }
+
             var company = await _context.Companies.FindAsync(id);
             _context.Companies.Remove(company);
             await _context.SaveChangesAsync();
